Fix wrong type for MessageBase.ID

+++ src/Helpers.TL.cs
@@ -87,25 +87,25 @@ namespace TL
 
 	partial class MessageBase
 	{
-		public abstract long ID { get; }
+		public abstract int ID { get; }
 		public abstract Peer Peer { get; }
 		public abstract DateTime Date { get; }
 	}
 	partial class MessageEmpty
 	{
-		public override long ID => id;
+		public override int ID => id;
 		public override Peer Peer => peer_id;
 		public override DateTime Date => default;
 	}
 	public partial class Message
 	{
-		public override long ID => id;
+		public override int ID => id;
 		public override Peer Peer => peer_id;
 		public override DateTime Date => date;
 	}
 	public partial class MessageService
 	{
-		public override long ID => id;
+		public override int ID => id;
 		public override Peer Peer => peer_id;
 		public override DateTime Date => date;
 	}
